+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 2021
 
+## 02/23 3.0.7
+* tweBBQ:
+  * Fix: `无汉化组`和`回复中无汉化组`的行为和预期不符
+  * Feat: 新增说明书回应，config新增字段`bbq.helpPage`
+  * Chore: 内部BBQ_ARGS的字段名字变更，由`no_group_info`、`no_group_info_in_reply`变为`no_logo`、`no_group_info`
+  
+
 ## 02/22 3.0.6
 * tweBBQ:
   * Feat: 查看未烤制内容队列